+++ spec/lib/danger/danger_core/environment_manager_spec.rb
@@ -56,6 +56,12 @@ RSpec.describe Danger::EnvironmentManager, use: :ci_helper do
       end
     end
 
+    it "loads Screwdriver" do
+      with_screwdriver_setup_and_is_a_pull_request do |system_env|
+        expect(described_class.local_ci_source(system_env)).to eq Danger::Screwdriver
+      end
+    end
+
     it "loads Semaphore" do
       with_semaphore_setup_and_is_a_pull_request do |system_env|
         expect(described_class.local_ci_source(system_env)).to eq Danger::Semaphore
@@ -145,6 +151,12 @@ RSpec.describe Danger::EnvironmentManager, use: :ci_helper do
     it "loads Local Git Repo" do
       with_localgitrepo_setup do |system_env|
         expect(described_class.pr?(system_env)).to eq(false)
+      end
+    end
+
+    it "loads Screwdriver" do
+      with_screwdriver_setup_and_is_a_pull_request do |system_env|
+        expect(described_class.pr?(system_env)).to eq(true)
       end
     end